varex_backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9
pydantic[email]==2.7.1
pydantic-settings==2.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
httpx==0.27.0
python-dotenv==1.0.0
boto3==1.34.0
slowapi==0.1.9
pdfminer.six==20231228
python-docx==1.1.0
redis==5.0.1
apscheduler==3.10.4
bleach==6.1.0
